Moved interface and class to seperate files.
