+++ src/coverlet.core/Instrumentation/IInstrumenter.cs
@@ -1,0 +1,9 @@
+namespace Coverlet.Core.Instrumentation
+{
+    internal interface IInstrumenter
+    {
+        bool CanInstrument();
+
+        InstrumenterResult Instrument();
+    }
+}
